결제 완료할때 포인트차감 및 업데이트, 쿠폰 제거 추가

+++ app/(user)/payments/success/page.tsx
@@ -15,11 +15,14 @@ import { cartService } from '@/services/cart';
 import Flex from '@/components/common/Flex';
 import { IPaymentRef, IPaymentSessionData } from '@/types/payment';
 import { couponService } from '@/services/coupon';
+import { myService } from '@/services/my';
+import { IUpdateUserRef } from '@/types/user';
 
 export default function PaymentSuccess() {
   const params = useSearchParams();
   const router = useRouter();
   const user = useUserStore((s) => s.user);
+  const { fetchUserData } = useUserStore();
 
   const [orderItems, setOrderItems] = useState<OrderItem[]>([]);
   const [paymentNumber, setPaymentNumber] = useState('');
@@ -187,7 +190,21 @@ export default function PaymentSuccess() {
         cartService.removesCart(cartIds);
         useCartStore.getState().removeStoreCarts(cartIds);
 
-        await couponService.deleteCoupon(selectedCoupon?.id ?? 0);
+        if (selectedCoupon?.id) {
+          await couponService.deleteCoupon(selectedCoupon?.id ?? 0);
+        }
+
+        if (user?.point) {
+          const calculPoint = user?.point - (paymentSessionData?.pointAmount ?? 0);
+
+          const userUpdatePoint: IUpdateUserRef = {
+            point: calculPoint,
+          };
+
+          await myService.updateUser(userUpdatePoint);
+
+          fetchUserData();
+        }
 
         // ⚓ paymentId + paymentNumber 확보
         const pid: number | null = paymentRes.data?.id ?? null;
@@ -215,6 +232,7 @@ export default function PaymentSuccess() {
     })();
   }, [
     discountAmount,
+    fetchUserData,
     orderItems,
     paymentAmount,
     paymentSessionData?.detailAddress,
